test(DeleteConfirmation): merge tests into one file
